EA-47390 (AIOOBE: StringBufferReplaceableByStringInspection$StringBufferReplaceableByStringFix.buildStringExpression)

+++ plugins/InspectionGadgets/InspectionGadgetsAnalysis/src/com/siyeh/ig/style/StringBufferReplaceableByStringInspection.java
@@ -128,6 +128,9 @@ public class StringBufferReplaceableByStringInspection extends BaseInspection {
       }
       final StringBuildingVisitor visitor = new StringBuildingVisitor(variable, stringExpression);
       codeBlock.accept(visitor);
+      if (visitor.hadProblem()) {
+        return;
+      }
       final List<PsiMethodCallExpression> expressions = visitor.getExpressions();
       variable.delete();
       for (int i = 0, size = expressions.size() - 1; i < size; i++) {
@@ -183,7 +186,10 @@ public class StringBufferReplaceableByStringInspection extends BaseInspection {
         else {
           final PsiExpressionList argumentList = methodCallExpression.getArgumentList();
           final PsiExpression[] arguments = argumentList.getExpressions();
-          if (arguments.length > 1) {
+          if (arguments.length == 0) {
+            return null;
+          }
+          else if (arguments.length > 1) {
             if (result.length() != 0) {
               result.append('+');
             }
@@ -237,9 +243,10 @@ public class StringBufferReplaceableByStringInspection extends BaseInspection {
 
     private static class StringBuildingVisitor extends JavaRecursiveElementVisitor {
 
-      private PsiVariable myVariable;
-      private StringBuilder myBuilder;
-      private List<PsiMethodCallExpression> expressions = new ArrayList();
+      private final PsiVariable myVariable;
+      private final StringBuilder myBuilder;
+      private final List<PsiMethodCallExpression> expressions = new ArrayList();
+      private boolean myProblem = false;
 
       public StringBuildingVisitor(@NotNull PsiVariable variable, StringBuilder builder) {
         myVariable = variable;
@@ -248,6 +255,9 @@ public class StringBufferReplaceableByStringInspection extends BaseInspection {
 
       @Override
       public void visitReferenceExpression(PsiReferenceExpression expression) {
+        if (myProblem) {
+          return;
+        }
         super.visitReferenceExpression(expression);
         if (expression.getQualifierExpression() != null) {
           return;
@@ -264,12 +274,18 @@ public class StringBufferReplaceableByStringInspection extends BaseInspection {
           parent = methodCallExpression.getParent();
           grandParent = parent.getParent();
         }
-        buildStringExpression(methodCallExpression, myBuilder);
+        if (buildStringExpression(methodCallExpression, myBuilder) == null) {
+          myProblem = true;
+        }
         expressions.add(methodCallExpression);
       }
 
       public List<PsiMethodCallExpression> getExpressions() {
         return expressions;
+      }
+
+      public boolean hadProblem() {
+        return myProblem;
       }
     }
   }
